+++ src/common/utils.c
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include <common/utils.h>
 #include <stddef.h>
 #include <stdio.h>
 #include <stdlib.h>
